Setting up leaderboard and admin pages
