+++ client/src/pages/admin/index.js
@@ -1,0 +1,1 @@
+export { Admin } from "./Admin";
