+++ client/src/pages/leaderboard/index.js
@@ -1,0 +1,1 @@
+export { Leaderboard } from "./Leaderboard";
